reconfigure structure and class to use IoT python SDK
