--- aws_greengrass_core_sdk/examples/Storyline/subscription_table.txt
@@ -1,72 +1,0 @@
-You can substitute the IDs and ARNs of the following and provide it through the CLI. This is provided in case copying from the PDF is not possible.
-
-
-aws greengrass create-subscription-definition-version --subscription-definition-id <SubscriptionId> --subscriptions \
-'[
-    {
-        "Id": "1",
-        "Source": "<Switch_Thing ARN>",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update",
-        "Target": "GGShadowService"
-    },
-    {
-        "Id": "2",
-        "Source": "<RobotArm_Thing ARN>",
-        "Subject": "/topic/state",
-        "Target": "<uptimeLambda ARN>"
-    },
-    {
-        "Id": "3",
-        "Source": "<RobotArm_Thing ARN>",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update",
-        "Target": "GGShadowService"
-    },
-    {
-        "Id": "4",
-        "Source": "GGShadowService",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update/delta",
-        "Target": "<RobotArm_Thing ARN>"
-    },
-    {
-        "Id": "5",
-        "Source": "<uptimeLambda ARN>",
-        "Subject": "/topic/metering",
-        "Target": "cloud"
-    },
-    {
-        "Id": "6",
-        "Source": "<messageLambda ARN>",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update",
-        "Target": "GGShadowService"
-    },
-    {
-        "Id": "7",
-        "Source": "cloud",
-        "Subject": "/topic/update",
-        "Target": "<messageLambda ARN>"
-    },
-    {
-        "Id": "8",
-        "Source": "GGShadowService",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update/accepted",
-        "Target": "<Switch_Thing ARN>"
-    },
-    {
-        "Id": "9",
-        "Source": "GGShadowService",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update/rejected",
-        "Target": "<Switch_Thing ARN>"
-    },
-    {
-        "Id": "10",
-        "Source": "GGShadowService",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update/accepted",
-        "Target": "<RobotArm_Thing ARN>"
-    },
-    {
-        "Id": "11",
-        "Source": "GGShadowService",
-        "Subject": "$aws/things/RobotArm_Thing/shadow/update/rejected",
-        "Target": "<RobotArm_Thing ARN>"
-    }
-]'
