--- aws_greengrass_core_sdk/manual/index.html
@@ -1,201 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    
-  
-  <!-- Licensed under the Apache 2.0 License -->
-  <link rel="stylesheet" type="text/css" href="https://boto3.readthedocs.io/en/latest/_static/fonts/open-sans/stylesheet.css" />
-  <!-- Licensed under the SIL Open Font License -->
-  <link rel="stylesheet" type="text/css" href="https://boto3.readthedocs.io/en/latest/_static/fonts/source-serif-pro/source-serif-pro.css" />
-  <link rel="stylesheet" type="text/css" href="https://boto3.readthedocs.io/en/latest/_static/css/bootstrap.min.css" />
-  <link rel="stylesheet" type="text/css" href="https://boto3.readthedocs.io/en/latest/_static/css/bootstrap-theme.min.css" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  
-    <title>Available Services &mdash; Boto 3 Docs 1.4.0 documentation</title>
-    
-    <link rel="stylesheet" href="https://boto3.readthedocs.io/en/latest/_static/guzzle.css" type="text/css" />
-    <link rel="stylesheet" href="https://boto3.readthedocs.io/en/latest/_static/pygments.css" type="text/css" />
-    <link rel="stylesheet" href="https://media.readthedocs.org/css/badge_only.css" type="text/css" />
-    
-    <script type="text/javascript">
-      var DOCUMENTATION_OPTIONS = {
-        URL_ROOT:    '../../',
-        VERSION:     '1.4.0',
-        COLLAPSE_INDEX: false,
-        FILE_SUFFIX: '.html',
-        HAS_SOURCE:  true
-      };
-    </script>
-    <script type="text/javascript" src="https://media.readthedocs.org/javascript/jquery/jquery-2.0.3.min.js"></script>
-    <script type="text/javascript" src="https://media.readthedocs.org/javascript/jquery/jquery-migrate-1.2.1.min.js"></script>
-    <script type="text/javascript" src="https://media.readthedocs.org/javascript/underscore.js"></script>
-    <script type="text/javascript" src="https://media.readthedocs.org/javascript/doctools.js"></script>
-    <script type="text/javascript" src="https://media.readthedocs.org/javascript/readthedocs-doc-embed.js"></script>
-    <link rel="top" title="Boto 3 Docs 1.4.0 documentation" href="../../index.html" />
-    <link rel="next" title="ACM" href="acm.html" />
-    <link rel="prev" title="S3" href="../../guide/s3.html" />
-  
-   
-
-  
-<!-- RTD Extra Head -->
-
-<!-- 
-Always link to the latest version, as canonical.
-http://docs.readthedocs.org/en/latest/canonical.html
--->
-<link rel="canonical" href="http://boto3.readthedocs.io/en/latest/reference/services/" />
-
-<link rel="stylesheet" href="https://media.readthedocs.org/css/readthedocs-doc-embed.css" type="text/css" />
-
-<script type="text/javascript" src="../../_static/readthedocs-data.js"></script>
-
-<!-- Add page-specific data, which must exist in the page js, not global -->
-<script type="text/javascript">
-READTHEDOCS_DATA['page'] = 'reference/services/index'
-</script>
-
-<script type="text/javascript" src="../../_static/readthedocs-dynamic-include.js"></script>
-
-<!-- end RTD <extrahead> --></head>
-  <body>
-    <div class="related">
-      <h3>Navigation</h3>
-      <ul>
-        <li class="right" style="margin-right: 10px">
-          <a href="../../genindex.html" title="General Index"
-             accesskey="I">index</a></li>
-        <li class="right" >
-          <a href="../../py-modindex.html" title="Python Module Index"
-             >modules</a> |</li>
-        <li class="right" >
-          <a href="acm.html" title="ACM"
-             accesskey="N">next</a> |</li>
-        <li class="right" >
-          <a href="../../guide/s3.html" title="S3"
-             accesskey="P">previous</a> |</li>
-        <li><a href="../../index.html">Greengrass Core SDK documentation</a> &raquo;</li> 
-      </ul>
-    </div>
-    <div class="container-wrapper">
-
-      <div id="mobile-toggle">
-        <a href="#"><span class="glyphicon glyphicon-align-justify" aria-hidden="true"></span></a>
-      </div>
-  <div id="left-column">
-    <div class="sphinxsidebar"><a href="
-    index.html" class="text-logo">AWS Greengrass Core SDK documentation</a><div class="sidebar-block">
-  <div class="sidebar-wrapper">
-    <h2>Table Of Contents</h2>
-  </div>
-  <div class="sidebar-toc">
-    
-    
-      <ul>
-<li class="toctree-l1"><a class="reference internal" href="quickstart.html">Quickstart</a></li>
-</ul>
-<ul class="current">
-<li class="toctree-l1 current"><a class="current reference internal" href="">Available Services</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="iot-data.html">IoTDataPlane</a></li>
-<li class="toctree-l2"><a class="reference internal" href="lambda.html">Lambda</a></li>
-</ul>
-</li>
-</ul>
-    
-  </div>
-</div>
-<div class="sidebar-block">
-  <div class="sidebar-wrapper">
-    <div id="main-search">
-      <form class="form-inline" action="../../search.html" method="GET" role="form">
-        <div class="input-group">
-          <input name="q" type="text" class="form-control" placeholder="Search...">
-        </div>
-        <input type="hidden" name="check_keywords" value="yes" />
-        <input type="hidden" name="area" value="default" />
-      </form>
-    </div>
-  </div>
-</div>
-      
-    </div>
-  </div>
-        <div id="right-column">
-          
-
-          <div role="navigation" aria-label="breadcrumbs navigation">
-            <ol class="breadcrumb">
-              <li><a href="index.html">Docs</a></li>
-              
-              <li>Available Services</li>
-            </ol>
-          </div>
-          
-<!--REGION_DISCLAIMER_DO_NOT_REMOVE-->
-          <div class="document clearer body">
-            
-  <div class="section" id="available-services">
-<h1>Available Services<a class="headerlink" href="#available-services" title="Permalink to this headline">¶</a></h1>
-<div class="toctree-wrapper compound">
-<li class="toctree-l1"><a class="reference internal" href="iot-data.html">IoTDataPlane</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="iot-data.html#client">Client</a></li>
-</ul>
-</li>
-<li class="toctree-l1"><a class="reference internal" href="lambda.html">Lambda</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="lambda.html#client">Client</a></li>
-</ul>
-</li>
-</div>
-</div>
-
-
-          </div>
-            
-  <div class="footer-relations">
-    
-    </div>
-    <div class="clearer"></div>
-  
-        </div>
-        <div class="clearfix"></div>
-    </div>
-    <div class="related">
-      <h3>Navigation</h3>
-      <ul>
-        <li class="right" >
-          <a href="../../py-modindex.html" title="Python Module Index"
-             >modules</a> |</li>
-        <li class="right" >
-          <a href="acm.html" title="ACM"
-             >next</a> |</li>
-        <li class="right" >
-          <a href="../../guide/s3.html" title="S3"
-             >previous</a> |</li>
-        <li><a href="../../index.html">Boto 3 Docs 1.4.0 documentation</a> &raquo;</li> 
-      </ul>
-    </div>
-
-<script type="text/javascript">
-  $("#mobile-toggle a").click(function () {
-    $("#left-column").toggle();
-  });
-</script>
-<script type="text/javascript" src="../../_static/js/bootstrap.js"></script>
-  <div class="footer">
-    &copy; Copyright 2014, Amazon.com, Inc.. Created manually.
-  </div>
-<script type="text/javascript" src="https://media.amazonwebservices.com/js/sitecatalyst/s_code.min.js"></script>
-<script type="text/javascript">
-s.prop66 = 'AWS SDK for Python (boto3)';
-s.eVar66 = 'D=c66';
-s.prop65 = 'API Reference';
-s.eVar65 = 'D=c65';
-var s_code = s.t();
-if (s_code) document.write(s_code);
-</script>
-  </body>
-</html>
