--- aws_greengrass_core_sdk/sdk/greengrass_common/__init__.py
@@ -1,6 +1,0 @@
-#
-# Copyright 2010-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-
-from .function_arn_fields import FunctionArnFields
-from .greengrass_message import GreengrassMessage
