--- aws_greengrass_core_sdk/sdk/greengrass_ipc_python_sdk/__init__.py
@@ -1,5 +1,0 @@
-#
-# Copyright 2010-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-
-from .ipc_client import IPCClient
