--- aws_greengrass_core_sdk/sdk/greengrasssdk/__init__.py
@@ -1,6 +1,0 @@
-#
-# Copyright 2010-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-
-from .client import client
-from .Lambda import StreamingBody
